requirements.txt
streamlit==1.33.0
pandas==2.2.2
matplotlib==3.8.4
scikit-learn==1.4.2
tensorflow==2.15.0
fpdf==1.7.2
python-dotenv==1.1.0
google-generativeai==0.8.5